Move catch.hpp to test/dependencies
