+++ test/cpp-dl/main.cpp
@@ -1,2 +1,2 @@
 #define CATCH_CONFIG_MAIN
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
