+++ test/cpp-raw/alloc.cpp
@@ -1,5 +1,5 @@
 #include "../cpp-utility/os.hpp"
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
 
 #include "testlib.hpp"
