+++ test/cpp-raw/interface_params.cpp
@@ -1,6 +1,6 @@
 
 #include "../cpp-utility/os.hpp"
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
 
 #define INTERCOM_FLATTEN_DECLARATIONS
 #include "testlib.hpp"
