+++ test/cpp-raw/iunknown.cpp
@@ -1,6 +1,6 @@
 
 #include "../cpp-utility/os.hpp"
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
 
 #include "testlib.hpp"
