+++ test/cpp-raw/main.cpp
@@ -1,4 +1,4 @@
 
 
 #define CATCH_CONFIG_MAIN
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
