+++ test/cpp-utility/msvc/os.cpp
@@ -3,7 +3,7 @@
 #include <mutex>
 #include <Objbase.h>
 
-#include "../catch.hpp"
+#include "../../dependencies/catch.hpp"
 #include "../os.hpp"
 #include "../generated/test_lib.hpp"
 #include "../../runpath/init.h"
