+++ test/cpp-wrapper/exceptions.cpp
@@ -3,7 +3,7 @@
 #include <functional>
 
 #include "../cpp-utility/os.hpp"
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
 
 #include "testlib.hpp"
 #include "../cpp-utility/dummy_interface.hpp"
