+++ test/cpp-wrapper/interface_wrappers.cpp
@@ -3,7 +3,7 @@
 #include <functional>
 
 #include "../cpp-utility/os.hpp"
-#include "../cpp-utility/catch.hpp"
+#include "../dependencies/catch.hpp"
 
 #define INTERCOM_FLATTEN_DECLARATIONS
 #include "testlib.hpp"
